Firefly/web_ui/requirements.txt
feedparser
flask-socketio
flask-login
Flask-SSLify